Added functionality for book borrowing/returning

+++ Controllers/BookController.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.StaticFiles;
 using System;
@@ -18,11 +19,15 @@ namespace WebLibrary.Controllers
     {
         private readonly MongoDBBookService _bookService;
         private readonly IMapper _mapper;
+        private readonly SignInManager<User> _signInManager;
+        private readonly UserManager<User> _userManager;
 
-        public BookController(MongoDBBookService bookService, IMapper mapper)
+        public BookController(MongoDBBookService bookService, IMapper mapper, SignInManager<User> signInManager, UserManager<User> userManager)
         {
             _bookService = bookService;
             _mapper = mapper;
+            _signInManager = signInManager;
+            _userManager = userManager;
         }
 
         [HttpGet("Index")]
@@ -90,7 +95,7 @@ namespace WebLibrary.Controllers
             bookDto.ImageUrl = book.ImageUrl;
             _mapper.Map(bookDto, book);
 
-            await _bookService.UpdatetAsync(bookId, book);
+            await _bookService.UpdatetAsync(book);
 
             return RedirectToAction(nameof(Index));
         }
@@ -98,10 +103,58 @@ namespace WebLibrary.Controllers
         [HttpPost("Delete")]
         public async Task<ActionResult> Delete(string bookId)
         {
-            /*add logic that denies deletion when any user has book*/
+            if ((await _bookService.FindByIdAsync(bookId)).AmountBorrowed <= 0)
+            {
+                return Ok(); // this needs to send warning message "Book cannot be deleted"
+            }
+            
             await _bookService.DeleteAsync(bookId);
 
             return RedirectToAction(nameof(Index));
+        }
+        
+        [HttpPost("Borrow")]
+        public async Task<ActionResult> Borrow(string bookId)
+        {
+            var book = await _bookService.FindByIdAsync(bookId);
+            var user = await _userManager.GetUserAsync(_signInManager.Context.User);
+
+            if (book.AmountAvalible <= 0 || user.BookIds.Contains(bookId) || user.BookIds.Count >= 6)
+            {
+                return Ok(); // this needs to send warning message "Book cannot be borrowed"
+            }
+
+            user.BookIds.Add(bookId);
+
+            book.AmountAvalible--;
+            book.AmountBorrowed++;
+
+            await _bookService.UpdatetAsync(book);
+            await _userManager.UpdateAsync(user);
+
+            return Json(book.AmountAvalible);
+        }
+
+        [HttpPost("Return")]
+        public async Task<ActionResult> Return(string bookId)
+        {
+            var book = await _bookService.FindByIdAsync(bookId);
+            var user = await _userManager.GetUserAsync(_signInManager.Context.User);
+
+            if (!user.BookIds.Contains(bookId))
+            {
+                return Ok(); // this needs to send warning message "Book cannot be returned"
+            }
+
+            user.BookIds.Remove(bookId);
+
+            book.AmountAvalible++;
+            book.AmountBorrowed--;
+
+            await _bookService.UpdatetAsync(book);
+            await _userManager.UpdateAsync(user);
+
+            return Json(book.AmountAvalible);
         }
 
         /* if needed in future for smth, can be just deleted tbh
